scripts/requirements-migrate.txt
# Uso: pip install -r scripts/requirements-migrate.txt
pymongo>=4.6
requests>=2.31
python-dotenv>=1.0
